Adicionadas melhorias nos códigos e comentários, para facilitar o estudo e entendimento

+++ P2/src/Recursividade/L11E5.java
@@ -10,8 +10,13 @@ public class L11E5 {
 		final String path = "src/Saidas/ResultadoEx5.txt";
 		final int x = 6;
 		try {
+			// Salva em branco só pra limpar o arquivo
 			GerenciadorDeArquivos.salvar(path, "");
-			int result = Util.resolveEImprimeEquacao(x,path);
+			// IMPORTANTE: Long foi usado porque 
+			// quando x é 6 ou mais, o número é 
+			// grande demais pra ser armazenado
+			// em variáveis "int"
+			long result = Util.resolveEImprimeEquacao(x,path);
 			System.out.println(result);
 			System.out.println("Arquivo atualizado!");
 		} catch (IOException e) {
